pyproject.toml
[tool.poetry]
name = "repl_python3_Resume"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto = "^2.49.0"
boto3 = "^1.20.23"
Flask = "^2.0.2"
replit = "^3.2.4"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
